split widgets from ui.c, add oc_ui_box_rect() and oc_ui_box_style()
